1. user테이블에서 토큰값이 저장되지 않는 문제 해결 2. 토큰값(id) 확인하고 있으면 기록된 user객체 리턴, 없으면 null 리턴 하는 걸로 수정했음
